+++ how-to/workspace-platform-starter/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v13
 
-- Renamed application to Workspace Platform Starter
+- Renamed application to Workspace Platform Starter (WPS)
 - Enabled strict mode
+- Added JSDoc Linting
 - App definition instanceMode extension. New options added: "new" (new means a new instance will be created if an intent is raised and the app specified but no instance id even if there are existing instances. The intent picker will also not show instances as the app provider has indicated they want to just launch a new instance).
 - If you app meta data indicates it supports the intent "OpenApp" and therefore fdc3.open then a new instance will be created if fdc3.open is called unless your app instance mode is "single". This is regardless of setting "multi" or "new" in instanceMode. If the intent "OpenApp" is called directly (not through fdc3.open) and your instanceMode is "multi" then it will prompt the user (through the intent resolver) to pick an instance or launch a new instance.